requirements.txt
requests
pandas
python-dotenv
openpyxl
sycopg2-binary
sqlalchemy